refactor(fetch): make fetch() stateless by using fresh Session per request
Each `fetch()` call now creates a new `Session` instance and closes it after
the request completes, preventing cookies and headers from leaking across
standalone calls. The underlying CurlMulti connection pool remains shared
via `getSharedMulti()`.

- Remove `shared-session.ts` module and inline its logic into `public.ts`
- Delete the `getFetchSession()` helper from `fetch.ts`
- Add test verifying cookies are not persisted between `fetch()` calls
